--- app/src/main/cpp/utils/AndroidOut.cpp
@@ -1,4 +1,0 @@
-#include "AndroidOut.h"
-
-AndroidOut androidOut("AO");
-std::ostream aout(&androidOut);
