+++ common/ndk_helper/build.gradle.kts
@@ -12,7 +12,7 @@ android {
 
         externalNativeBuild {
             cmake {
-                cppFlags += "-std=c++14 -frtti -Wall -Werror"
+                cppFlags += "-std=c++11 -frtti -Wall -Werror"
                 targets += "ndk_helper"
                 arguments(
                     "-DNDK_HELPER_PATH=${project(":common:ndk_helper").projectDir}",
