+++ common/ndk_helper/src/main/cpp/NDKHelper.hpp
@@ -4,5 +4,6 @@
 #include "Model_Impl.hpp"
 #include "manager/TimeManager.hpp"
 #include "manager/AssetManager.hpp"
+#include "utils/AndroidOut.hpp"
 
 #endif //MESH_ROTATOR_NDKHELPER_HPP
