+++ common/ndk_helper/src/main/cpp/test/Model_ImplTest.cpp
@@ -1,7 +1,7 @@
 #include "../Model_Impl.hpp"
 #include "mock/MeshMock.hpp"
+#include "../utils/Utility.hpp"
 #include <gtest/gtest.h>
-#include <memory>
 
 namespace ndk_helper {
     namespace test {
@@ -12,7 +12,7 @@ namespace ndk_helper {
             void SetUp() override {
                 mesh::Vertex vertex{{0.0f, 1.0f, 2.0f}};
                 mesh::Mesh mesh{{vertex}, {}};
-                model_ = std::make_unique<mdl::Model_Impl>(
+                model_ = utils::make_unique<mdl::Model_Impl>(
                     mdl::Model_Impl(
                         mesh,
                         glm::vec3(0.0f),
