+++ common/ndk_helper/src/main/cpp/utils/AndroidOut.cpp
@@ -1,0 +1,4 @@
+#include "AndroidOut.hpp"
+
+ndk_helper::AndroidOut androidOut("AO");
+std::ostream aout(&androidOut);
